consolidated instance and patcher state
