+++ src/reducers/index.ts
@@ -6,7 +6,6 @@ import { editor } from "./editor";
 import { instances } from "./instances";
 import { graph } from "./graph";
 import { nofitications } from "./notifications";
-import { patchers } from "./patchers";
 import { settings } from "./settings";
 import { sets } from "./sets";
 import { transport } from "./transport";
@@ -18,7 +17,6 @@ export const rootReducer = combineReducers({
 	instances,
 	graph,
 	nofitications,
-	patchers,
 	settings,
 	sets,
 	transport
