+++ src/selectors/instances.ts
@@ -1,10 +1,40 @@
-import { Map as ImmuMap } from "immutable";
+import { Map as ImmuMap, Seq } from "immutable";
 import { RootStateType } from "../lib/store";
 import { PatcherInstanceRecord } from "../models/instance";
 import { createSelector } from "reselect";
 import { getPatcherIdsByIndex } from "./graph";
 import { ParameterRecord } from "../models/parameter";
 import { MessagePortRecord } from "../models/messageport";
+import { PatcherRecord } from "../models/patcher";
+import { SortOrder } from "../lib/constants";
+
+export const getPatchers = (state: RootStateType): ImmuMap<PatcherRecord["id"], PatcherRecord> => {
+	return state.instances.patchers;
+};
+
+export const getPatcher = createSelector(
+	[
+		getPatchers,
+		(state: RootStateType, name: string): string => name
+	],
+	(patchers, name): PatcherRecord | undefined => {
+		return patchers.get(name);
+	}
+);
+
+const collator = new Intl.Collator("en-US");
+export const getPatchersSortedByName = createSelector(
+	[
+		getPatchers,
+		(state: RootStateType, order: SortOrder): SortOrder => order
+	],
+	(patchers, order): Seq.Indexed<PatcherRecord> => {
+		return patchers.valueSeq().sort((pA, pB) => {
+			return collator.compare(pA.name.toLowerCase(), pB.name.toLowerCase()) * (order === SortOrder.Asc ? 1 : -1);
+		});
+	}
+);
+
 
 export const getInstances = (state: RootStateType): ImmuMap<PatcherInstanceRecord["id"], PatcherInstanceRecord> => state.instances.instances;
